% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsverse.R
\docType{package}
\name{phsverse}
\alias{phsverse}
\title{\code{phsverse} package}
\description{
It was created with \href{https://github.com/jdtrat/metamakr}{metamakr} and
combines our PHS packages on GitHub,
\href{https://github.com/Public-Health-Scotland/phstemplates}{phstemplates},
\href{https://github.com/Public-Health-Scotland/phsmethods}{phsmethods} and
\href{https://github.com/Public-Health-Scotland/phsstyles}{phsstyles}.
}
\details{
See the README on
\href{https://github.com/Public-Health-Scotland/phsverse#readme}{GitHub}.
}
